code improvements from testmerge

+++ code/game/objects/effects/decals/cleanable.dm
@@ -94,7 +94,8 @@
 
 /obj/effect/decal/cleanable/wash(clean_types)
 	..()
-	qdel(src)
+	if(clean_types in list(CLEAN_WASH, CLEAN_SCRUB))
+		qdel(src)
 	return TRUE
 
 /obj/effect/decal/cleanable/proc/can_bloodcrawl_in()
